collection-primitive-list

+++ src/main/java/hello1/koddata/utils/collection/primitive/lists/IntList.java
@@ -2,30 +2,30 @@ package hello1.koddata.utils.collection.primitive.lists;
 
 import hello1.koddata.utils.math.MathUtils;
 
-public class ShortList {
+public class IntList {
 
-    private short[] data;
+    private int[] data;
     private int size = 0;
     private static final int DEFAULT_CAPACITY = 8;
 
-    public ShortList() {
-        data = new short[DEFAULT_CAPACITY];
+    public IntList() {
+        data = new int[DEFAULT_CAPACITY];
     }
 
-    public ShortList(int capacity){
+    public IntList(int capacity){
         if(capacity <= 0){
             throw new IllegalArgumentException("Capacity must be greater than zero");
         }
 
-        data = new short[capacity];
+        data = new int[capacity];
     }
 
-    public void add(short i){
+    public void add(int i){
         ensureCapacity();
         data[size++] = i;
     }
 
-    public void add(int index, short i){
+    public void add(int index, int i){
         if(index < 0 || index > size){
             throw new IndexOutOfBoundsException("Index: " + index);
         }
@@ -37,11 +37,11 @@ public class ShortList {
         size++;
     }
 
-    public void addAll(short[] array, int count) {
+    public void addAll(int[] array, int count) {
         int requiredCapacity = size + count;
         if (requiredCapacity > data.length) {
             int newCapacity = MathUtils.nearestPowerOf2(requiredCapacity * 2);
-            short[] newData = new short[newCapacity];
+            int[] newData = new int[newCapacity];
             System.arraycopy(data, 0, newData, 0, size);
             data = newData;
         }
@@ -49,23 +49,23 @@ public class ShortList {
         size += count;
     }
 
-    public void addAll(short[] array){
+    public void addAll(int[] array){
         addAll(array, array.length);
     }
 
-    public short get(int index){
+    public int get(int index){
         checkIndex(index);
         return data[index];
     }
 
-    public void set(int index, short i){
+    public void set(int index, int i){
         checkIndex(index);
         data[index] = i;
     }
 
-    public short remove(int index){
+    public int remove(int index){
         checkIndex(index);
-        short removed = data[index];
+        int removed = data[index];
         for(int i = index; i < size - 1; i++){
             data[i] = data[i + 1];
         }
@@ -87,7 +87,7 @@ public class ShortList {
     }
 
     public void removeAll(){
-        data = new short[DEFAULT_CAPACITY];
+        data = new int[DEFAULT_CAPACITY];
         size = 0;
     }
 
@@ -96,7 +96,7 @@ public class ShortList {
         if (currentCapacity <= DEFAULT_CAPACITY) return;
         if (size <= currentCapacity / 4) {
             int newCapacity = Math.max(DEFAULT_CAPACITY, currentCapacity / 2);
-            short[] newData = new short[newCapacity];
+            int[] newData = new int[newCapacity];
             System.arraycopy(data, 0, newData, 0, size);
             data = newData;
         }
@@ -105,7 +105,7 @@ public class ShortList {
     private void ensureCapacity() {
         if(size == data.length){
             int newCapacity = data.length * 2;
-            short[] newInt = new short[newCapacity];
+            int[] newInt = new int[newCapacity];
             System.arraycopy(data, 0, newInt, 0, data.length);
             data = newInt;
         }
